requirements.txt
streamlit==1.29.0
langchain==0.0.325
openai==0.28.1
python-dotenv==1.0.0
chromadb==0.4.15
tiktoken==0.5.2
PyPDF2==3.0.1
python-docx==1.1.0
numpy<2
packaging<24
protobuf<5
rich<14
tenacity<9
anyio<4
importlib-metadata<7
ollama==0.5.3
sentence-transformers==5.1.0.0
langchain==0.0.325
openai==0.28.1
python-dotenv==1.0.0
chromadb==0.4.15
tiktoken==0.5.2
PyPDF2==3.0.1
python-docx==1.1.0
numpy<2
packaging<24
protobuf<5
rich<14
tenacity<9
anyio<4
importlib-metadata<7
